requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
exceptiongroup==1.3.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.0
langchain-core==1.1.0
langchain-huggingface==1.1.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.49
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0
